refactor: relocate composer autocomplete to shared AutocompletePopup
