+++ apps/meteor/client/components/AutocompletePopup/hooks/useComposerBoxPopup.ts
@@ -4,8 +4,8 @@ import type { MutableRefObject } from 'react';
 import { useEffect, useCallback, useState, useRef } from 'react';
 
 import { useComposerBoxPopupQueries } from './useComposerBoxPopupQueries';
-import { useChat } from '../../contexts/ChatContext';
-import type { ComposerPopupOption } from '../../contexts/ComposerPopupContext';
+import { useChat } from '../../../views/room/contexts/ChatContext';
+import type { ComposerPopupOption } from '../ComposerPopupOption';
 
 type ComposerBoxPopupImperativeCommands<T> = MutableRefObject<
 	| {
