+++ apps/meteor/client/components/AutocompletePopup/hooks/useComposerBoxPopupQueries.ts
@@ -3,8 +3,8 @@ import { keepPreviousData, useQueries } from '@tanstack/react-query';
 import { useEffect, useState } from 'react';
 
 import { useEnablePopupPreview } from './useEnablePopupPreview';
-import { slashCommands } from '../../../../../app/utils/client/slashCommand';
-import type { ComposerPopupOption } from '../../contexts/ComposerPopupContext';
+import { slashCommands } from '../../../../app/utils/client/slashCommand';
+import type { ComposerPopupOption } from '../ComposerPopupOption';
 
 export const useComposerBoxPopupQueries = <T extends { _id: string; sort?: number }>(filter: unknown, popup?: ComposerPopupOption<T>) => {
 	const [counter, setCounter] = useState(0);
