+++ apps/meteor/client/components/AutocompletePopup/hooks/useEnablePopupPreview.ts
@@ -1,4 +1,4 @@
-import type { ComposerPopupOption } from '../../contexts/ComposerPopupContext';
+import type { ComposerPopupOption } from '../ComposerPopupOption';
 
 export const useEnablePopupPreview = <T extends { _id: string; sort?: number }>(filter: unknown, popup?: ComposerPopupOption<T>) =>
 	popup && !popup.preview && (popup?.triggerLength ? typeof filter === 'string' && popup.triggerLength - 1 < filter.length : true);
