+++ apps/meteor/client/views/room/composer/ComposerBoxPopupPreview.tsx
@@ -4,7 +4,7 @@ import { useEndpoint } from '@rocket.chat/ui-contexts';
 import type { ForwardedRef, ReactNode } from 'react';
 import { forwardRef, useEffect, useId, useImperativeHandle } from 'react';
 
-import type { ComposerBoxPopupProps } from './ComposerBoxPopup';
+import type { ComposerBoxPopupProps } from '../../../components/AutocompletePopup/ComposerBoxPopup';
 import { useChat } from '../contexts/ChatContext';
 
 type ComposerBoxPopupPreviewItem = { _id: string; type: 'image' | 'video' | 'audio' | 'text' | 'other'; value: string; sort?: number };
